2.0.1 - A few changes
* Remove python3.7 support
* Remove python3.7 unit test hacks
* Aggregate vcorelib.math imports
* Add DEFAULT_DEPTH symbol
