+++ local/variables/package.yaml
@@ -1,4 +1,4 @@
 ---
 major: 2
 minor: 0
-patch: 0
+patch: 1
