+++ tests/io/archive/test_make_archive.py
@@ -9,7 +9,7 @@ import shutil
 from tempfile import TemporaryDirectory
 
 # internal
-from tests.resources import get_archives_root, skip_archive
+from tests.resources import get_archives_root
 
 # module under test
 from vcorelib.io.archive import make_archive
@@ -33,8 +33,6 @@ def test_make_archive():
             # Test expected successes.
             assert make_archive(src)[0] is not None
             for archive in ["tar", "tar.bz2", "tar.lzma", "tar.xz", "zip"]:
-                if skip_archive(archive):
-                    continue
                 assert make_archive(src, archive)[0] is not None
 
             with TemporaryDirectory() as tmpdir_new:
