+++ vcorelib/__init__.py
@@ -1,7 +1,7 @@
 # =====================================
 # generator=datazen
 # version=3.1.2
-# hash=4f5520f60ed64fc76ac05b90d573bee8
+# hash=1b16823d3c8b9c309d235f1ecd2e33d2
 # =====================================
 
 """
@@ -10,7 +10,7 @@ Useful defaults and other package metadata.
 
 DESCRIPTION = "A collection of core Python utilities."
 PKG_NAME = "vcorelib"
-VERSION = "2.0.0"
+VERSION = "2.0.1"
 
 # vcorelib-specific content.
 DEFAULT_INDENT = 2
