+++ vcorelib/math/analysis/__init__.py
@@ -1,0 +1,16 @@
+"""
+A module for aggregating symbols for import.
+"""
+
+# internal
+from vcorelib.math.analysis.average import DEFAULT_DEPTH, MovingAverage
+from vcorelib.math.analysis.rate import RateTracker, ns_to_s
+from vcorelib.math.analysis.rate.limiter import RateLimiter
+
+__all__ = [
+    "MovingAverage",
+    "DEFAULT_DEPTH",
+    "ns_to_s",
+    "RateTracker",
+    "RateLimiter",
+]
